requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
tzdata==2022.6
djangorestframework~=3.14.0
environ~=1.0
python-decouple~=3.6
drf-yasg==1.21.4

django-filter~=22.1
django-ckeditor~=6.5.1